requirements.txt
requests==2.17.1
newsapi-python==0.2.3